I added a README.md file to help people navigate the blog I added another blog into the database. I edited new_blog.sh to let people use their favorite EDITOR.

+++ blog/new_blog.sh
@@ -48,24 +48,47 @@ while [[ -z $NAME ]]; do
 	read NAME
 done
 
-HOOK=""
-LINE="placeholder"
-NL='\n'
-echo "Enter the summary/hook (when you are done, enter a blank line) (also, this program will expand any escaped characters):"
-read LINE
-HOOK=$LINE
-while [[ ! -z $LINE ]]; do
-	read LINE
-	if [[ -z $LINE ]]; then
-		break
-	fi
-	HOOK=$HOOK$NL$LINE
-done
+#HOOK=""
+#LINE="placeholder"
+#NL='\n'
+#echo "Enter the summary/hook (when you are done, enter a blank line) (also, this program will expand any escaped characters):"
+#read LINE
+#HOOK=$LINE
+#while [[ ! -z $LINE ]]; do
+#	read LINE
+#	if [[ -z $LINE ]]; then
+#		break
+#	fi
+#	HOOK=$HOOK$NL$LINE
+#done
 
-sed -i '1 d' $DATABASE # Delete first line in file 
+HOOK=""
+HOOKF=/tmp/new_blog_hook.txt
+touch $HOOKF
+if [[ ! -e $HOOKF ]]; then
+	echo "$HOOKF was unable to be created. Safely exiting."
+	exit
+fi
+echo "Write the hook in this file. This file will disappear after you submit it. (this line is not ignored by the parser)" > $HOOKF
+$EDITOR $HOOKF
+HOOK=$(cat $HOOKF)
+rm $HOOKF
+
+
+
 TMP=/tmp/new_blog_prepend.txt
-DUP=/tmp/new_blog_database_duplicate.txt
 touch $TMP
+if [[ ! -e $TMP ]]; then
+	echo "$TMP was unable to be created. Safely exiting."
+	exit
+fi
+DUP=/tmp/new_blog_database_duplicate.txt
+touch $DUP
+if [[ ! -e $DUP ]]; then
+	echo "$DUP was unable to be created. Safely exiting."
+	exit
+fi
+sed -i '1 d' $DATABASE # Delete first line in database 
 cp $DATABASE $DUP
 
 echo -e '{' >> $TMP
